backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
orjson==3.9.10

# DB
SQLAlchemy==2.0.25
pyodbc==5.1.0

# Data science
pandas==2.2.0
numpy==1.26.3
lightgbm==4.2.0
scikit-learn==1.4.0
shap==0.44.1
joblib==1.3.2